.apm/skills/wiki/scripts/requirements.txt
# Requirements for wiki generation scripts
PyYAML>=6.0
